requirements.txt
# general module
django>3.2,<3.3
python-decouple
pillow
djangorestframework

# third party modules
markdown
django-filter
drf-yasg[validation]

# deployment modules

# Email third party modules
django-mail-templated

# refactoring and test
black
flake8

# Documentations
# coreapi

#jwt
# djangorestframework-simplejwt